[stack-manager] Move configureStack logic into StackProjectImporter
