+++ src/com/haskforce/haskell/project/externalSystem/stack/StackOpenProjectProvider.scala
@@ -1,13 +1,7 @@
 package com.haskforce.haskell.project.externalSystem.stack
 
-import java.io.File
-
-import com.haskforce.settings.HaskellBuildSettings
-import com.haskforce.utils.ExecUtil
 import com.intellij.openapi.externalSystem.importing.AbstractOpenProjectProvider
-import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.util.ThrowableComputable
 import com.intellij.openapi.vfs.VirtualFile
 
 class StackOpenProjectProvider extends AbstractOpenProjectProvider {
@@ -21,52 +15,6 @@ class StackOpenProjectProvider extends AbstractOpenProjectProvider {
     projectDirectory: String,
     project: Project
   ): Unit = {
-    if (!configureStack(projectDirectory, project)) return
     StackProjectImporter.importProject(project, projectDirectory)
-  }
-
-  /**
-   * Attempt ot configure the stack exe and stack.yaml for the given project.
-   */
-  private def configureStack(
-    projectDirectory: String,
-    project: Project
-  ): Boolean = {
-    val s = HaskellBuildSettings.getInstance(project)
-    val stackYamlFile = new File(projectDirectory, "stack.yaml")
-    if (!stackYamlFile.isFile) return false
-    val stackYamlPath = stackYamlFile.getCanonicalPath
-    // When the default stack exe is not valid, attempt to implicitly
-    // configure it. If we can't get a valid stack exe, bail out.
-    if (!currentStackExeIsValid(s)
-          && !implicitlyConfigureStackExe(project, s)) {
-      return false
-    }
-    s.setStackFile(stackYamlPath)
-    s.setUseStack(true)
-    true
-  }
-
-  private def currentStackExeIsValid(s: HaskellBuildSettings): Boolean = {
-    Option(s.getStackPath).exists(p => new File(p).canExecute)
-  }
-
-  private def implicitlyConfigureStackExe(
-    project: Project,
-    s: HaskellBuildSettings
-  ): Boolean = {
-    val go: ThrowableComputable[Option[String], Exception] = () => {
-      Option(ExecUtil.locateExecutableByGuessing("stack"))
-    }
-    // Guard against 'Synchronous execution on EDT' errors.
-    ProgressManager.getInstance().runProcessWithProgressSynchronously(
-      go,
-      "Locating stack executable",
-      true,
-      project
-    ) match {
-      case None => false
-      case Some(p) => s.setStackPath(p); true
-    }
   }
 }
