+++ src/com/haskforce/haskell/project/externalSystem/stack/StackProjectImporter.scala
@@ -1,13 +1,23 @@
 package com.haskforce.haskell.project.externalSystem.stack
 
+import java.io.File
+
+import com.haskforce.HaskellModuleType
+import com.haskforce.codeInsight.visibleModules.GhcPkgDumpProjectDependencyCache
+import com.haskforce.settings.HaskellBuildSettings
+import com.haskforce.tooling.ghcPkg.GhcPkgDumpProjectCacheService
+import com.haskforce.utils.ExecUtil
+import com.intellij.execution.configurations.GeneralCommandLine
 import com.intellij.openapi.externalSystem.importing.ImportSpecBuilder
 import com.intellij.openapi.externalSystem.model.DataNode
 import com.intellij.openapi.externalSystem.model.project.ProjectData
 import com.intellij.openapi.externalSystem.service.execution.ProgressExecutionMode
-import com.intellij.openapi.externalSystem.service.project.{ExternalProjectRefreshCallback, ProjectDataManager}
 import com.intellij.openapi.externalSystem.service.project.manage.ExternalProjectsManagerImpl
+import com.intellij.openapi.externalSystem.service.project.{ExternalProjectRefreshCallback, ProjectDataManager}
 import com.intellij.openapi.externalSystem.util.ExternalSystemUtil
+import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.ThrowableComputable
 
 object StackProjectImporter {
 
@@ -15,6 +25,7 @@ object StackProjectImporter {
     project: Project,
     projectDirectory: String
   ): Unit = {
+    if (!configureStack(projectDirectory, project)) return
     ExternalProjectsManagerImpl.getInstance(project).runWhenInitialized(() => {
       ExternalSystemUtil.ensureToolWindowInitialized(
         project,
@@ -50,10 +61,57 @@ object StackProjectImporter {
     override def onSuccess(externalProject: DataNode[ProjectData]): Unit = {
       if (externalProject == null) return
       importData(externalProject)
+      GhcPkgDumpProjectDependencyCache.getInstance(project).reload()
+      ()
     }
 
     private def importData(externalProject: DataNode[ProjectData]): Unit = {
       ProjectDataManager.getInstance().importData(externalProject, project, false)
+    }
+  }
+
+  /**
+   * Attempt ot configure the stack exe and stack.yaml for the given project.
+   */
+  private def configureStack(
+    projectDirectory: String,
+    project: Project
+  ): Boolean = {
+    val s = HaskellBuildSettings.getInstance(project)
+    val stackYamlFile = new File(projectDirectory, "stack.yaml")
+    if (!stackYamlFile.isFile) return false
+    val stackYamlPath = stackYamlFile.getCanonicalPath
+    // When the default stack exe is not valid, attempt to implicitly
+    // configure it. If we can't get a valid stack exe, bail out.
+    if (!currentStackExeIsValid(s)
+      && !implicitlyConfigureStackExe(project, s)) {
+      return false
+    }
+    s.setStackFile(stackYamlPath)
+    s.setUseStack(true)
+    true
+  }
+
+  private def currentStackExeIsValid(s: HaskellBuildSettings): Boolean = {
+    Option(s.getStackPath).exists(p => new File(p).canExecute)
+  }
+
+  private def implicitlyConfigureStackExe(
+    project: Project,
+    s: HaskellBuildSettings
+  ): Boolean = {
+    val go: ThrowableComputable[Option[String], Exception] = () => {
+      Option(ExecUtil.locateExecutableByGuessing("stack"))
+    }
+    // Guard against 'Synchronous execution on EDT' errors.
+    ProgressManager.getInstance().runProcessWithProgressSynchronously(
+      go,
+      "Locating stack executable",
+      true,
+      project
+    ) match {
+      case None => false
+      case Some(p) => s.setStackPath(p); true
     }
   }
 }
